// src/app/dashboard/modules/inbox/modules/app-message-list/enum/message.enum.ts
export enum EMailProcess {
  REMINDER = 'reminder',
  NEW_SHARED_MAILBOX = 'newSharedMailBox'
}

export enum EPopupConversionTask {
  SELECT_OPTION = 'SELECT_OPTION',
  CREATE_TASK = 'CREATE_TASK',
  MOVE_TASK = 'MOVE_TASK'
}

export enum EAppMessageCreateType {
  NewAppMessage = 'NEW_APP_MESSAGE',
  NewAppMessageDone = 'NEW_APP_MESSAGE_DONE',
  ReplyViaApp = 'REPLY_VIA_APP'
}

export enum EDeleteInLineType {
  CONVERSATION_LINKED_APP = 'CONVERSATION_LINKED_APP',
  DRAFT_APP_MESSAGE = 'DRAFT_APP_MESSAGE'
}