// src/app/task-detail/enums/task-detail.enum.ts
export enum EViewDetailMode {
  INTERNAL_NOTE = 'internal-note',
  CONVERSATION = 'conversation',
  TASK = 'task',
  MESSAGE = 'message',
  EMAIL = 'email',
  REMINDER = 'reminder',
  APP_MESSAGE = 'app-messages',
  SMS_MESSAGE = 'sms-messages'
}

export enum ECreatePopupOptionType {
  CREATE_NEW_TASK = 'CREATE_NEW_TASK',
  BULK_CREATE_TASKS = 'BULK_CREATE_TASKS'
}